% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{falrec}
\alias{falrec}
\title{falrec}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5968 rows and 5 columns.
}
\usage{
falrec
}
\description{
falrec
}
\keyword{datasets}
